Format with black

+++ scripts/chapter_list_creator.py
@@ -73,13 +73,13 @@ for filename in os.listdir(collations_path):
         # Adding the manuscript id
         verse_ids_in_this_manuscript["manuscript_id"] = manuscript_id
         dfs.append(verse_ids_in_this_manuscript)
-        #print(verse_ids_in_this_manuscript)
+        # print(verse_ids_in_this_manuscript)
 
 master_df = pd.concat(dfs)
 
-master_df['verse'] = pd.to_numeric(master_df['verse'], errors='coerce')
-master_df.dropna(subset=['verse'], inplace=True)
-master_df['verse'] = master_df['verse'].astype(int)
+master_df["verse"] = pd.to_numeric(master_df["verse"], errors="coerce")
+master_df.dropna(subset=["verse"], inplace=True)
+master_df["verse"] = master_df["verse"].astype(int)
 
 master_df = master_df.sort_values(by=["book", "chapter", "verse", "manuscript_id"])
 
@@ -126,29 +126,35 @@ book_dict = {
 }
 
 
-relation_for_apparatus = master_df[['book', 'chapter', 'verse', 'manuscript_id']]
-relation_for_apparatus['book_name'] =  relation_for_apparatus['book'].replace(book_dict)
+relation_for_apparatus = master_df[["book", "chapter", "verse", "manuscript_id"]]
+relation_for_apparatus["book_name"] = relation_for_apparatus["book"].replace(book_dict)
 
 dictionary_for_apparatus = {}
-for book_name in relation_for_apparatus['book_name'].unique():
+for book_name in relation_for_apparatus["book_name"].unique():
     dictionary_for_apparatus[book_name] = {}
-    this_book = relation_for_apparatus[relation_for_apparatus['book_name']==book_name]
-    for chapter in this_book['chapter'].unique():
+    this_book = relation_for_apparatus[relation_for_apparatus["book_name"] == book_name]
+    for chapter in this_book["chapter"].unique():
         chapter = int(chapter)
         dictionary_for_apparatus[book_name][chapter] = {}
-        this_chapter = this_book[this_book['chapter']==chapter]
-        for verse in this_chapter['verse'].unique():
+        this_chapter = this_book[this_book["chapter"] == chapter]
+        for verse in this_chapter["verse"].unique():
             verse = int(verse)
-            this_verse = this_chapter[this_chapter['verse']==verse]
-            manuscripts_for_this_verse = this_verse[['verse', 'manuscript_id']].groupby('verse')['manuscript_id'].apply(list)
-            dictionary_for_apparatus[book_name][chapter][verse] = manuscripts_for_this_verse.iloc[0]
-            
+            this_verse = this_chapter[this_chapter["verse"] == verse]
+            manuscripts_for_this_verse = (
+                this_verse[["verse", "manuscript_id"]]
+                .groupby("verse")["manuscript_id"]
+                .apply(list)
+            )
+            dictionary_for_apparatus[book_name][chapter][
+                verse
+            ] = manuscripts_for_this_verse.iloc[0]
+
 # Specify the file path for the JSON file
-file_path = '../apparatus/manuscript_verse_relation.json'
+file_path = "../apparatus/manuscript_verse_relation.json"
 
 # Open the file in write mode and use json.dump() to write the dictionary to the file with pretty formatting
 
-with open(file_path, 'w') as f:
+with open(file_path, "w") as f:
     json.dump(dictionary_for_apparatus, f, indent=4)
 
 ############################################
@@ -157,7 +163,11 @@ with open(file_path, 'w') as f:
 ############################################
 ############################################
 
-master_df = master_df.drop(columns=['verse']).drop_duplicates().sort_values(by=['manuscript_id'])
+master_df = (
+    master_df.drop(columns=["verse"])
+    .drop_duplicates()
+    .sort_values(by=["manuscript_id"])
+)
 
 # Does not include verse information
 grouped_df = (
